reimplement child packet recursion, now save ;)
G2PacketSerializer.c
	implement a function to handle the half decoded packet from
	extract_from_stream. Here you get your child packets one after
	the other.
	Implement another function to half decode those child packets.
	goto 1
G2PacketSerializer.h
	export new function
G2Packet.c
	use new functions and build/use framework to recurse child packets

Target reached!
As beautiful as it was, that you could decode a complete packet only
by the packet format, with all childs an recieve a "ready to use" tree,
as malicius was it.
Bad packets with high recursion of NOP/not understanded childs would
waste your CPU and memory to generate a tree, you simply would throw away.

Now the recursion is coupled to the "understanding" of the packet.
Mamory allocation take place on one location (an can be optimized out)

Only drawback: This must mature and be refoctored, so finaly we can work
with minimal allocations/copys. Rework of Complete chain from reciveing 
of bytes to every handler...

Makefile
	fix dependencies
G2MainServer.c
	make time longer