change 6bit hack to two times 7bit
after a read up on valid packet chars it is clear, it wount fit in
6bit. Lower characters are mandated for private packets. On the other
hand numbers are a clear part of names (-0x30 aka '0' is the limit,
leaving out some interpunctation which could be desired).
Everything else is also likely so basicaly we can only exclude control
characters (<0x20) as a check for desync and 8bit ASCII (>0x7F)(until
some frenchmen start to send localized packets...).

 G2PacketTyperGenerator.c
	generate differant header and change checks
 G2Packet.c
	change typer

expect change to cleancheck in Serializer.