change vsnprintf to use function jump table and strchrnul
Since the compiler generates a jumptable for a big switch we can put
our functions directly in a jump table and dynamicaly decode the format
string from function to function (decode-as-we-go).
And again, going bytewise over the format is silly, use strchrnul to find
the % or the nul-byte. We do not generate a super duper special func for
this.

 lib/generic/strchrnul.c
 lib/ppc/strchrnul.c
 lib/strchrnul.c
 lib/x86/strchrnul.c
	strchrnul for non-GNU platforms
 lib/vsnprintf.c
	use it
	new decoding

 lib/ppc/memand.c
 lib/ppc/memneg.c
 lib/ppc/memxor.c
 lib/ppc/strlen.c
 lib/ppc/strncasecmp_a.c
 lib/ppc/strnlen.c
 lib/ppc/strnpcpy.c
 lib/ppc/mem_searchrn.c
	use cache hint loads
 lib/x86/x86.h
	change the cache hints. This hurts benchmarks, but helps real apps
	which also have other (more important) stuff in the cache.