make MMX usage a little saver
When using MMX, one garbels the FPU stack. Any usage of the fpu
as fpu without "cleaning up" the MMX mess would result in crashes.
That was one of the main problems of MMX.
Luckily not for us, fpu usage is in general not allowed in g2cd
(OK, i'm not the kernel, so can not control everything and esp.
libs, but you get the idea...).
But we now have a legit exception: reading/scalling the "float"
coordinates for the profile and GPS packets.
But since we could have used MMX several times before we come here
(at least on old cpus, memcpy, strlen), we would crash.
Fix by emmiting one "emms" before the offending float foo sequence,
so we should be good to go.

 lib/my_bitops.c
 lib/my_bitops.h
 lib/x86/my_bitops.c
	add a little wrapper to emit an emms
 G2MainServer.c
	call it