add libcurl support
From time to time we need to connect to a Gnutella Web Cache (GWC) by
http protocol. Doing a smiple http connection by hand is quite simple.

The problem is that http has a lot of complexity up it's sleeve.
(encoding, compression, redirection, proxys, foo).

For that reason the GWC specification says a _proper__ http client
has to be used, not some quick hacked up version, to prevent harm
for the GWCs from broken/useless accesses.

libcurl is a everything and the kitchensink bloat, but a proper
http client.
It's use is _highly_ recommended.


 configure.ac
	test for libcurl
 G2KHL.c
	use libcurl
 README
	add some text about libcurl