Fix neasty from OS X
OS X likes to report OOB data unsolicited in it's kqueue implementation.
This could send you into an infinite loop when you receive a single TCP
packet it the URG-Bit, because your appp gets woken by the kernel again
and again and again, but you do not know to look for OOB data, if you
do not expect it. (and no kqueue impl. among the BSDs does that).

The main code already handles such a case by killing connections which
are in a state we don't know.

Give that code the info from the kqueue wrapper that something is amies.

G2ConHelper.c
 add debug text for the case
G2Handler.c
 commentos
lib/my_epoll_kqueue.c
 pass flag along