change x86 jmp offset calc
WARNING!

This change will probably BREAK older systems.

Somehow binutils now decides to calculate the expression changed
differently.
So we jump into nirvana.
The + 3 is the fix to get a sane result with this new way (leaving
out the -1f totally f'es it up).

This used to work, so i smell a change in binutils and expect
this change will break older versions.

Unfortunatly i do not even know what to goole for...