silence "may be used uninitialised"
our new gcc 5 overflow checking mull fallback is a little hard to grasp
for older compiler.
Make sure to set res in any case, so the compiler shuts up.

This is unescessery, because if the mull "failed" the result should not
be used (and is not, the compilers flow analysis foo was just to weak),
but should prevent problems and the compiler should remove it again.